requirements.txt
streamlit==1.31.0
google-generativeai==0.3.2
langchain==0.1.0
langchain-google-genai==0.0.6
beautifulsoup4==4.12.3
python-dotenv==1.0.1
requests==2.31.0
duckduckgo-search==4.1.1
aiohttp==3.9.1
nest-asyncio==1.5.8
